backend/requirements.txt
# ============================================
# Django Core
# ============================================
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# ============================================
# Database
# ============================================
pymongo==4.6.1
sqlparse==0.4.4
dnspython==2.4.2

# ============================================
# Authentication & Security
# ============================================
django-cors-headers==4.3.1
PyJWT==2.8.0
cryptography==42.0.0

# ============================================
# Celery & Redis
# ============================================
celery==5.3.6
redis==5.0.1
django-redis==5.4.0

# ============================================
# File Storage
# ============================================
cloudinary==1.38.0
django-cloudinary-storage==0.3.0
Pillow==10.2.0

# ============================================
# Email
# ============================================
django-templated-mail==1.1.1

# ============================================
# API Documentation
# ============================================
drf-spectacular==0.27.1

# ============================================
# Utilities
# ============================================
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1

# ============================================
# PDF Generation
# ============================================
reportlab==4.0.9
pypdf==4.0.1

# ============================================
# Validation & Serialization
# ============================================
marshmallow==3.20.2

# ============================================
# Rate Limiting
# ============================================
django-ratelimit==4.1.0

# ============================================
# Testing
# ============================================
pytest==8.0.0
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
faker==22.6.0

# ============================================
# Code Quality
# ============================================
black==24.1.1
flake8==7.0.0
isort==5.13.2
pylint==3.0.3
mypy==1.8.0

# ============================================
# Development Tools
# ============================================
django-extensions==3.2.3
ipython==8.20.0
django-debug-toolbar==4.2.0

# ============================================
# WSGI Server (Production)
# ============================================
gunicorn==21.2.0
uvicorn[standard]==0.27.0
